Split event manager controller into five admin controllers
