+++ convergence_games/apps/frontend/admin/controllers/__init__.py
@@ -1,3 +1,13 @@
-from ._event_manager import EventManagerController
+from ._allocation import AllocationController
+from ._players import PlayersController
+from ._schedule import ScheduleController
+from ._settings import SettingsController
+from ._submissions import SubmissionsController
 
-__all__ = ["EventManagerController"]
+__all__ = [
+    "AllocationController",
+    "PlayersController",
+    "ScheduleController",
+    "SettingsController",
+    "SubmissionsController",
+]
